requirements.txt
flask
openai-whisper
yt-dlp